cleanup and move classes around
