+++ src/main/java/net/kyori/registry/impl/map/IdMapImpl.java
@@ -21,21 +21,34 @@
  * OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE
  * SOFTWARE.
  */
-package net.kyori.registry.api;
+package net.kyori.registry.impl.map;
 
+import it.unimi.dsi.fastutil.ints.Int2ObjectMap;
+import it.unimi.dsi.fastutil.objects.Object2IntMap;
 import org.checkerframework.checker.nullness.qual.NonNull;
 
+import java.util.OptionalInt;
+import java.util.function.IntPredicate;
+
 /**
- * A read-only component of an id registry with a default key and value.
+ * A simple implementation of an id map.
  *
  * @param <V> the value type
  */
-public interface DefaultedRegistryIdentifier<V> {
-    /**
-     * Gets the id for {@code value}.
-     *
-     * @param value the value
-     * @return the id
-     */
-    int idOrDefault(final @NonNull V value);
+public class IdMapImpl<V> extends AbstractIdMap<V> {
+    private final IntPredicate empty;
+
+    public IdMapImpl(final @NonNull Int2ObjectMap<V> idToV, final @NonNull Object2IntMap<V> vToId, final @NonNull IntPredicate empty) {
+        super(idToV, vToId);
+        this.empty = empty;
+    }
+
+    @Override
+    public @NonNull OptionalInt id(final @NonNull V value) {
+        final int id = this.vToId.getInt(value);
+        if (!this.empty.test(id)) {
+            return OptionalInt.of(id);
+        }
+        return OptionalInt.empty();
+    }
 }
