+++ src/test/java/net/kyori/registry/map/IncrementalIdMapImplTest.java
@@ -21,34 +21,31 @@
  * OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE
  * SOFTWARE.
  */
-package net.kyori.registry.id.map;
+package net.kyori.registry.map;
 
 import it.unimi.dsi.fastutil.ints.Int2ObjectOpenHashMap;
 import it.unimi.dsi.fastutil.objects.Object2IntOpenHashMap;
-import net.kyori.registry.api.map.IdMap;
+import net.kyori.registry.api.map.IncrementalIdMap;
 import org.junit.jupiter.api.Test;
 
-import static org.junit.jupiter.api.Assertions.*;
+import static org.junit.jupiter.api.Assertions.assertEquals;
 
-class IdMapImplTest {
+class IncrementalIdMapImplTest {
     private static final int DEFAULT = -1000;
-    private final IdMap<String> map = IdMap.create(new Int2ObjectOpenHashMap<>(), new Object2IntOpenHashMap<String>() {
+    private final IncrementalIdMap<String> map = IncrementalIdMap.create(new Int2ObjectOpenHashMap<>(), new Object2IntOpenHashMap<String>() {
         {
             this.defaultReturnValue(DEFAULT);
         }
     }, value -> value == -1000);
 
     @Test
-    void testId() {
-        assertFalse(this.map.id("foo").isPresent());
-        this.map.put(32, "foo");
-        assertEquals(32, this.map.id("foo").orElse(DEFAULT));
-    }
-
-    @Test
-    void testGet() {
-        assertNull(this.map.get(32));
-        this.map.put(32, "foo");
-        assertEquals("foo", this.map.get(32));
+    void testPut() {
+        final int i0 = this.map.put("abc");
+        assertEquals(0, i0);
+        assertEquals("abc", this.map.get(i0));
+        final int i1 = this.map.put("def");
+        assertEquals(1, i1);
+        assertEquals("def", this.map.get(i1));
+        assertEquals("abc", this.map.get(i0));
     }
 }
