+++ src/test/java/net/kyori/registry/registry/bidirectional/BidirectionalRegistryTest.java
@@ -21,10 +21,10 @@
  * OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE
  * SOFTWARE.
  */
-package net.kyori.registry;
+package net.kyori.registry.registry.bidirectional;
 
 import com.google.common.collect.ImmutableList;
-import net.kyori.registry.api.Registry;
+import net.kyori.registry.api.registry.Registry;
 import org.junit.jupiter.api.Test;
 
 import java.util.Iterator;
@@ -32,7 +32,7 @@ import java.util.Iterator;
 import static com.google.common.truth.Truth.assertThat;
 import static org.junit.jupiter.api.Assertions.*;
 
-class BiRegistryTest {
+class BidirectionalRegistryTest {
     @Test
     void testGetKey() {
         final Registry<String, String> registry = Registry.create();
