+++ src/test/java/net/kyori/registry/registry/bidirectional/DefaultValueBidirectionalRegistryTest.java
@@ -21,40 +21,27 @@
  * OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE
  * SOFTWARE.
  */
-package net.kyori.registry.api.map;
+package net.kyori.registry.registry.bidirectional;
 
-import it.unimi.dsi.fastutil.ints.Int2ObjectMap;
-import it.unimi.dsi.fastutil.objects.Object2IntMap;
-import org.checkerframework.checker.nullness.qual.NonNull;
-import org.checkerframework.checker.nullness.qual.Nullable;
+import net.kyori.registry.api.registry.Registry;
+import net.kyori.registry.impl.registry.DefaultValueRegistry;
+import org.junit.jupiter.api.Test;
 
-/**
- * An abstract implementation of an id map.
- *
- * @param <V> the value type
- */
-public abstract class AbstractIdMap<V> implements IdMap<V> {
-    protected final Int2ObjectMap<V> idToV;
-    protected final Object2IntMap<V> vToId;
+import static org.junit.jupiter.api.Assertions.*;
 
-    protected AbstractIdMap(final @NonNull Int2ObjectMap<V> idToV, final @NonNull Object2IntMap<V> vToId) {
-        this.idToV = idToV;
-        this.vToId = vToId;
-    }
+class DefaultValueBidirectionalRegistryTest {
+    @Test
+    void testDefaultKeyValue() {
+        final DefaultValueRegistry<String, String> container = new DefaultValueRegistry<>(Registry.create(), "_default");
+        assertEquals("_default", container.defaultKey());
 
-    @Override
-    public final @NonNull V put(final int id, final @NonNull V value) {
-        this.put0(id, value);
-        return value;
-    }
+        assertNull(container.get("aaa"));
+        assertNull(container.get("ccc"));
 
-    protected void put0(final int id, final @NonNull V value) {
-        this.idToV.put(id, value);
-        this.vToId.put(value, id);
-    }
+        container.register("_default", "bbb");
+        assertEquals("bbb", container.get("aaa"));
 
-    @Override
-    public @Nullable V get(final int id) {
-        return this.idToV.get(id);
+        assertNotNull(container.get("ccc"));
+        assertEquals("bbb", container.getOrDefault("ccc"));
     }
 }
